+++ src/Broadcast/Broadcast.php
@@ -50,10 +50,11 @@ trait Broadcast
      * containing a Progress object for all broadcasts currently in-progress.
      *
      * @param array $messages The messages to send: an array of arrays, containing parameters to pass to messages.sendMessage.
+     * @param bool $pin Whether to also pin the last sent message.
      */
-    public function broadcastMessages(array $messages, ?Filter $filter = null): int
+    public function broadcastMessages(array $messages, ?Filter $filter = null, bool $pin = false): int
     {
-        return $this->broadcastCustom(new ActionSend($this, $messages), $filter);
+        return $this->broadcastCustom(new ActionSend($this, $messages, $pin), $filter);
     }
     /**
      * Forwards a list of messages to all peers (users, chats, channels) of the bot.
@@ -70,10 +71,11 @@ trait Broadcast
      * @param mixed $from_peer Bot API ID or Update, from where to forward the messages.
      * @param list<int> $message_ids IDs of the messages to forward.
      * @param bool $drop_author If true, will forward messages without quoting the original author.
+     * @param bool $pin Whether to also pin the last sent message.
      */
-    public function broadcastForwardMessages(mixed $from_peer, array $message_ids, bool $drop_author = false, ?Filter $filter = null): int
+    public function broadcastForwardMessages(mixed $from_peer, array $message_ids, bool $drop_author = false, ?Filter $filter = null, bool $pin = false): int
     {
-        return $this->broadcastCustom(new ActionForward($this, $this->getID($from_peer), $message_ids, $drop_author), $filter);
+        return $this->broadcastCustom(new ActionForward($this, $this->getID($from_peer), $message_ids, $drop_author, $pin), $filter);
     }
 
     /**