+++ src/InternalDoc.php
@@ -253,10 +253,11 @@ abstract class InternalDoc
      * @param mixed $from_peer Bot API ID or Update, from where to forward the messages.
      * @param list<int> $message_ids IDs of the messages to forward.
      * @param bool $drop_author If true, will forward messages without quoting the original author.
+     * @param bool $pin Whether to also pin the last sent message.
      */
-    public function broadcastForwardMessages(mixed $from_peer, array $message_ids, bool $drop_author = false, ?\danog\MadelineProto\Broadcast\Filter $filter = null): int
+    public function broadcastForwardMessages(mixed $from_peer, array $message_ids, bool $drop_author = false, ?\danog\MadelineProto\Broadcast\Filter $filter = null, bool $pin = false): int
     {
-        return $this->wrapper->getAPI()->broadcastForwardMessages($from_peer, $message_ids, $drop_author, $filter);
+        return $this->wrapper->getAPI()->broadcastForwardMessages($from_peer, $message_ids, $drop_author, $filter, $pin);
     }
     /**
      * Sends a list of messages to all peers (users, chats, channels) of the bot.
@@ -273,10 +274,11 @@ abstract class InternalDoc
      * containing a Progress object for all broadcasts currently in-progress.
      *
      * @param array $messages The messages to send: an array of arrays, containing parameters to pass to messages.sendMessage.
+     * @param bool $pin Whether to also pin the last sent message.
      */
-    public function broadcastMessages(array $messages, ?\danog\MadelineProto\Broadcast\Filter $filter = null): int
+    public function broadcastMessages(array $messages, ?\danog\MadelineProto\Broadcast\Filter $filter = null, bool $pin = false): int
     {
-        return $this->wrapper->getAPI()->broadcastMessages($messages, $filter);
+        return $this->wrapper->getAPI()->broadcastMessages($messages, $filter, $pin);
     }
     /**
      * Convert generator, promise or any other value to a promise.