+++ src/Stream/Common/HashedBufferedStream.php
@@ -128,7 +128,7 @@ final class HashedBufferedStream implements BufferedProxyStreamInterface, Buffer
      */
     public function setExtra($hash): void
     {
-        $rev = \strpos($hash, '_rev');
+        $rev = \str_contains($hash, '_rev');
         $this->rev = false;
         if ($rev !== false) {
             $hash = \substr($hash, 0, $rev);